Added conversion of targets to robot's point of view in feet

+++ targets.hpp
@@ -21,7 +21,7 @@ namespace acv {
 typedef struct {
   std::string type = ""; /**< either robot, ball, or some other target */
   std::string color = ""; /**< either red or blue */
-  int coords[3]; /**< coordinates in 3-space */
+  float coords[3]; /**< coordinates in 3-space */
 } Target;
 
 class Targets
